Fixed bug in searching and no slash in pointnames
This is a breaking change in terms of output files of the code.
The data file generated by this project now would not keep the prefix
of \ anymore. This creates unnecessary headache in dealing with the
data file. Thus if you update the package, teh pointnames would change
from what they were and you might need to fetch fresh copy of data
again. That could again cause downstream issues in your code. This
needed to be done though.

Rolling minor version due to this breaking feature.